requirements.txt
yfinance
pandas
numpy
python-telegram-bot